+++ src/structures/enter_service_get_type.rs
@@ -22,7 +22,7 @@ impl EnterServiceGetType {
             errors.push(OcppError::FieldValidationError {
                 field: "id".to_string(),
                 source: vec![OcppError::FieldCardinalityError {
-                    cardinality: self.id.len() as i32,
+                    cardinality: self.id.len(),
                     lower: 0,
                     upper: 36,
                 }],
